refactor: rename "AbstractEntity" package into "core"
